pyproject.toml
[tool.poetry]
name = "face-sentiment-agent"
version = "0.1.0"
description = "AI-driven agent to detect face sentiment and suggest some good read, video etc."
authors = ["Manmohan Negi <manmohan.hnbgu@gmail.com>"]
license = "Apache License 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9, <3.13"
python-dotenv = ">=1.0.0"
Pillow = ">=10.0.0"
opencv-python = ">=4.9.0.80"
langchain = "^0.2.0"
langchain-google-genai = "^1.0.5"
langchain-community = "^0.2.0"
tavily-python = ">=0.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
black = "^24.4.0"
google-adk = { version = "^1.0.0", extras = ["eval"] }
google-genai = "<1.25.0"
pytest-asyncio = "^0.23.7"

[tool.poetry.group.deployment]
optional = true

[tool.poetry.group.deployment.dependencies]
absl-py = "^2.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
